requirements.txt
websockets>=10.4
numpy>=1.20.0
PyYAML>=6.0
faster-whisper>=0.9.0
soundfile>=0.12.1
librosa>=0.10.0
torch>=2.0.0 